is_use_source true run pass
